Refactor: Delete Home component and add its functionality to App.jsx
